WML theme from Joe Buckshin <jbuckshin(a)gmail>
This patch adds a wml theme to mythweb.  This enables mythweb to be
used with most phone browsers that are not html capable (ie, openwave
4.x browsers).

The browser detection code was tweaked to use the wml theme when a wap
browser is detected and html content is not accepted.

An additional change to recorded_programs.php was made to support wml
specific redirects when the wml theme is in use.

* converted time/date formats to language specific formats (generic_date, generic_time)
* program listings reworked for usability
* more use of language strings
* fixed paging error on recording schedule
* changed php short open tags to long format
* tab to spaces cleanup

Up for debate -- do we even still need the WAP theme?


git-svn-id: http://svn.mythtv.org/svn/trunk@4704 7dbf422c-18fa-0310-86e9-fd20926502f2