References #1133. Card input refactoring; bumps MYTH_PROTO_VERSION.
The V4L Channel class enumated the inputs available on a card differently from the other channel classes. This unifies that enumeration, which allows us to put more of the support for inputs in the ChannelBase class. I also changed the name of the currentcapchannel variable to 'currentInputID' which should be less confusing to people writing new recorder (this confused Jafa when he was writing the hdhomerun recorder).

This also helps with #1133 because we enumerate the inputs based on the inputid in all the channel classes, so an input on a seperate piece of hardware can be added to the list (though we don't do that yet in this commit).

This also changes the protocol for input switching by replacing "TOGGLE_INPUTS" with three new commands "GET_CONNECTED_INPUTS", "GET_INPUT" and "SET_INPUT". This means we can add menu item for each inpt that allows you to switch to a particular input, but the immediate improvement is that we can ignore the request to switch inputs when there is only one input on the card. Though my longer term plan is to implement a 'switch channel lineup' key rather than the current 'Y' and 'C' commands, which presume that you know which card and input on that card that you are currently on.



git-svn-id: http://svn.mythtv.org/svn/trunk@8973 7dbf422c-18fa-0310-86e9-fd20926502f2