== IMPORTANT == This commit removes the .htaccess file used by MythWeb.  The file is replaced by an apache.conf file that provides the same features (plus a few more).  Your old .htaccess files should still work, but subversion might decide to delete them for you.  There is nothing I can do about this, and I apologize.
This also creates an INSTALL file with detailed installation instructions, and updates the README accordingly.

I've had enough complaints that the .htaccess file doesn't always work, and have had to suggest creating an apache conf file, anyway, that it's hopefully easier to put all of this information into a single file.  This also opens the way for using lighttpd, which I have had a handful of requests for.

git-svn-id: http://svn.mythtv.org/svn/trunk@12155 7dbf422c-18fa-0310-86e9-fd20926502f2