This fixes the out-of-sync previews on different frontends bug.
This introduces a new MythEvent UPDATE_PROG_INFO which tells the frontends to update a particular program info. This is only used for bookmark updates in this commit, but can be used to eliminate many instances of the RECORDING_LIST_CHANGE event (to speed up the Watch Recordings screen).

This also expands the mainserver MESSAGE command to forward the extra data in events not just the message. Hence the protocol update.

Finally, this adds a bookmarkupdate timestamp to the recorded table. This allows us to avoid many of our preview regens since we are no longer using the lastmodified column as a proxy for bookmark updates.

Make sure to recompile the plugins and all backends/frontends.

git-svn-id: http://svn.mythtv.org/svn/trunk@22932 7dbf422c-18fa-0310-86e9-fd20926502f2