Fix a few places in MythWeb affected by the ProgramInfo refactor.  Refs #8406.
This simplifies Program.php, and fixes the recorded.php page (which also parses some ProgramInfo data).  Initial patch by Daniel K, with some modifications by me.  I removed the attempt to reuse the Program object's parsing of ProgramInfo data because I couldn't get it to work.  That can be modified later, in a separate change, but this at least gets the page working, again (the same way it worked before).


git-svn-id: http://svn.mythtv.org/svn/trunk@24703 7dbf422c-18fa-0310-86e9-fd20926502f2