MythWeb: Show Season and Episode in the recorded page.
Displays the Season and Episode of a recorded program now that they're available (when mythmetadatalookup is called) for recordings.  if you haven't already done lookups in master for your recording rules to get the inetrefs in there, and to enable the mythmetadatalookup job, you should!  You can even turn the job on in mythweb's rule editor! :)

commit OK'd by Rob Smith, who graciously offered to fix the episode number padding when he gets a chance.