Allow metadata lookup in Mythweb.
I've been using this patch from Doug Haber since the day he submitted
it, never with any issue.  Adds a metadata lookup button to recording
rules screens, and recordings.  Also adds a link to the TVDB homepage
for the episode and a popup akin to the MythVideo titel selection popup.

Hooks into the backend Services API to perform lookups.

Fixes #9939.