Split the x-pixmap off from channelicon in clean_blue
Seems that the channelicon tweaks in a5bc3484f56a0ad9ddfd332bacba8594d171fc8c
forced the pixmap size (only in clean_blue skin) to 70x70 max, which borked
the aspect ratio.  Split the CSS to have two separate items, and with
x-pixmap back the way it was previously.