New Italian MythWeb category translation file
This is *not* a full translation, only the category file is currently translated.

I am not adding it as a selectable language using the English translation as was
proposed by the author as this would only confuse users.

I have tried contacting the author (to offer possible solutions) more than a week
ago but have received no reply.

Closes #10328

Signed-off-by: Nicolas Riendeau <nriendeau@mythtv.org>