Fix overriding of manual recordings in mythweb
Turns out we were creating slightly incorrect records for the record table.
Comparing to how it works from the frontend, I found that we were not
copying over the "search" field from the original manual recording rule, so
the override never matched up with it.  To fix that, I removed the line that
overrided "search" to 0 before saving.

Fixes #9343