Fix MythWeb Upcoming Recordings programinfo usage.
The programinfo refactor in #8406 changed the order of fields returned
in a programinfo string list, and a reference was missed in the MythWeb
Schedule class.  The result is that any recording rule with recordid of
6 was skipped/didn't show in Upcoming Recordings.

The original intent of the code was to skip recordings with recstatus of
rsNotListed (as shown in commits e728aea5 and 508103511 , committed for
which has since been modified, again, by a41e9657.  (Therefore, the
index in this commit is correct for 0.25-fixes and above, but should be
18 for 0.24-fixes.)

Thanks to JK on the -users list for finding the issue.