Fix MythWeb Upcoming Recordings programinfo usage.
The programinfo refactor in #8406 changed the order of fields returned
in a programinfo string list, and a reference was missed in the MythWeb
Schedule class.  The result is that any recording rule with recordid of
6 was skipped/didn't show in Upcoming Recordings.

The original intent of the code was to skip recordings with recstatus of
rsNotListed (as shown in commits e728aea5 and 508103511 , committed for

Thanks to JK on the -users list for finding the issue.

(Note that this is basically a backport of the same fix in 39a5dcd10
(master) and 89a347c852 (fixes/0.25), but with the proper index for
fixes/0.24, which does not contain a41e9657/was not reordered a second
time.)