Set php/pl connection timezone to +0:00
The PHP and Perl code uses unix UTC timestamps for internal timekeeping
and use mysql timezone conversions to convert UTC time to DB local time.
Since we keep the DB in UTC now this was doing an unwanted conversion.
Luckily turning this off is just a matter of setting a connection
time_zone value.