Reduce cache headers for program details ajax popup.
As it stands, these are being set to 10 days beyond the program start
time.  This means that mythweb will often show stale data for programs
that have been updated since they were initially added to the guide.

This reduces the cache time to 2 hours, which seems like a reasonable
time to cache for, yet still get timely updates.