Refs #12752 - [mythweb] Avoid double quoting sql
$db->escape already wraps the result in 's so there is no need to do it in the
caller.

This is a little complicated in searchMusic since the %'s need to be inside the
quoting but we do not want them to be escaped themselves. Concat them with the
search term using a SQL CONCAT().

Signed-off-by: Stuart Auchterlonie <stuarta@mythtv.org>
(cherry picked from commit d2469d2e3978a1288f79157990cf015d421d95d4)