Fixes #13191 - Make mythweb use of CDN optional.
Adds a new database setting 'mythweb_use_cdn' which defaults
to true. If you wish to use local copies of the libraries or
run on an isolated network you will need to set this to false

There is no gui in the mythweb setting page to configure this
at this time. Patches for this will be accepted.

In addition this add the updated libraries locally.