Fix warning messages thrown from php 7.3.
Most of these are simple errors where a variable hasn't been defined
as an array and thus be counted, or a constant hasn't been defined and
therefore can't be used.

Also fixes a couple of 'invalid range in character class' warnings by
quoting the dash in the bracketed character class to prevent
interpreting it as a range of characters.